made ImageModal and Gallery work with main site and made them more versatile

+++ src/components/Image.svelte
@@ -4,6 +4,7 @@
 	export let generatedImage: string;
 	export let generatedImageAlt: string;
 	export let id: number;
+	export let canDelete: boolean;
 
 	async function deleteImage(id: number) {
 		const payload = {
@@ -31,8 +32,10 @@
 		src={generatedImage}
 		alt={generatedImageAlt}
 	/>
-	<button
-		on:click|stopPropagation={() => deleteImage(id)}
-		class="absolute hidden w-6 h-6 bg-[url(/close.png)] bg-cover right-1 top-1 bg-[rgba(0,0,0,0.1)] opacity-75 group-hover:block hover:opacity-100 hover:bg-[rgba(0,0,0,0.25)]"
-	/>
+	{#if canDelete}
+		<button
+			on:click|stopPropagation={() => deleteImage(id)}
+			class="absolute hidden w-6 h-6 bg-[url(/close.png)] bg-cover right-1 top-1 bg-[rgba(0,0,0,0.1)] opacity-75 group-hover:block hover:opacity-100 hover:bg-[rgba(37,36,36,0.25)]"
+		/>
+	{/if}
 </button>
